CRCs are 32 bits.
git-svn-id: svn://svn.h5l.se/heimdal/trunk/heimdal@1418 ec53bebd-3082-4978-b11e-865c3cabbd6b

+++ lib/krb5/decrypt.c
@@ -11,7 +11,7 @@ krb5_decrypt (krb5_context context,
 	      krb5_data *result)
 {
   u_char *p = (u_char *)ptr;
-  u_long my_crc, her_crc;
+  u_int32_t my_crc, her_crc;
   des_cblock key;
   des_key_schedule schedule;
